refactor: SocialInfo Interface 적용 #11
