+++ src/domain/member/member.module.ts
@@ -1,12 +1,12 @@
-import { MemberService } from './member.service';
 import { Module } from '@nestjs/common';
 import { MemberController } from './member.controller';
+import { MemberService } from './member.service';
 import { MemberRepository } from './member.repository';
 
 @Module({
   imports: [],
   controllers: [MemberController],
   providers: [MemberService, MemberRepository],
-  exports: [MemberRepository],
+  exports: [MemberService],
 })
 export class MemberModule {}
